fix(m6b): #82 外部评审 AI3(Devin)处置(P2×2 全修 + P3 七改三录)
评审基准=e46a7c5(当前 HEAD,无时间差),结论「可 merge,无 P1」。全项处置:

P2(全修):
- **P2-1 dump 错误路径泄漏 inflight 记账**:note_inflight(2*bgra) 之后有
  6 个失败出口(CreateBitmap/cast/CopyFromRenderTarget/Map/pitch 守卫/
  Unmap)无一复位,卡住的值直接进紧随的关窗统计行——与 AI1 P2-4 在
  upload_tile 修的同类、未同步到 dump 臂。修法:readback 尾段拆成
  dump_readback 内方法,调用侧双路径统一 note_inflight(0)
- **P2-2 gpu<=cap 断言比实现保证的界严**:Tiles→Base 缩放过渡不反向清
  tile LRU(有意缓存),巨图 1:1 填满 LRU 后缩到 fit 会报 cap+overview>
  cap(S5c 全量扫描是侥幸绿)。修法:Base{level>=1} 臂把 LRU trim 到
  cap-base 的新 headroom(新 Lru::trim_to:逐冷驱逐、不改插入帽),严格界
  现在真成立;账本范围说明同步改写

P3(七改):
- **missing 双重计数**(e46a7c5 我引入):上传循环逐块 +1 与末尾核验相加,
  极端打「2N of N」。改为单一 post-pass 真值(上传失败/拒绝/同帧驱逐统一
  覆盖),删除逐块死增量
- **死参数清除**:工作集预算改动后 resident 已不参与计划决策——plan_frame
  与 tile_requests 的 resident 参数、gpu.rs 的闭包、测试的 &never/&hit 全
  删(文档同步:驻留不再是计划输入,绘制期 pre-touch 才是执行点)
- downscale_box 累加器 u32→u64(零成本防未来块几何)
- 退化 fringe .max(1) 注释改写(邻块共享整数边本就覆盖该列;规则是防御性
  所有权保证,代价 ≤1px 重叠——非「防洞」)
- -tile 无条件吞词的契约注释(与 -dump-viewport 同款,-tile photo.png
  吃掉文件名)
- prepare 失败路径 scene 复位(Err 出口不再让关窗统计行报上一成功帧)
- cap 单次采样说明(驱动压力再采样钩子=Lru::set_cap,在案)

P3(三录,处置表说明理由):O(n²)(仅强制诊断帧可感,HashMap 化留 #90)/
Tiles↔Base 全量重传 base(账目诚实的既定代价)/连续上传失败升级(留 #90
连同失败链重构)

验证:cargo test 576/0;clippy -D warnings / fmt 绿;smoke82 46/0/0、
smoke80 28/0/1、smoke81 72/0/0(golden 全 match)

+++ src/mip.rs
@@ -80,14 +80,16 @@ pub(crate) fn downscale_box(src: &[u8], image_w: i32, image_h: i32, level: u32) 
         for dx in 0..dest_w as usize {
             let x0 = dx * sw / dest_w as usize;
             let x1 = ((dx + 1) * sw / dest_w as usize).max(x0 + 1).min(sw);
-            let mut sum = [0u32; 4];
-            let mut count = 0u32;
+            // u64 accumulators: zero cost, and immune to any future block
+            // geometry (external review AI3 P3).
+            let mut sum = [0u64; 4];
+            let mut count = 0u64;
             for y in y0..y1 {
                 let row = y * sw * 4;
                 for x in x0..x1 {
                     let px = row + x * 4;
                     for (c, s) in sum.iter_mut().zip(&src[px..px + 4]) {
-                        *c += u32::from(*s);
+                        *c += u64::from(*s);
                     }
                     count += 1;
                 }
